ci: Add capslock to track dependency capabilities
Adds capslock to the CI pipeline. A snapshot is made of the `node/`
directory and stored in version control. When a new dependency is added,
the snapshot should be regenerated.
This tool should help us improve supply-chain security by preventing
packages which have increased their capabilities since our last snapshot
from being committed to the repository.
Any dependencies with increased privileges must be committed manually
along with a new snapshot.