node: Add transfer-verifier proof-of-concept binary
Adds a sub command under `guardiand` that can be used to do extra
integrity checks when messages are emitted from the core bridge.

This script reads from the EVM Watcher's RPC connection and monitors
LogMessagePublished events. When one is detected, and the sender is the
token bridge, the program parses the transaction receipt to ensures that
a Transfer() or Deposit() event was emitted in the same transaction.
This is an invariant of the Token Bridge. It logs an error when this
does not occur.