node/tv: Fix amount calculations
The previous implementation of verifying the amounts looked for strict
equality. This often isn't the case as the Token Bridge (de)normalizes
amounts based on the token's decimals which results in a different
value from the amount present in the payload
This commit adds a function that uses a heuristic to check if the amount
is likely normalized. If so, it will consider it alongside the
destination as usual and from there mark whether the transfer 'looks
normal'
This PR adds simple unit tests for this heuristic function though it
does not check errors (invalid arguments), only whether the function
correctly returns true or false where appropriate.